# contracts/openapi/ibms-protected-api-v1.openapi.yaml
openapi: 3.0.3
info:
  title: VANTARIS IBMS Protected API
  version: ibms.contract.v1
  description: |
    Draft protected API boundary for VANTARIS IBMS.
    Sources: SECURITY-A6, A7, A8B, A10, CONTRACTS-A1, CORE-A0.
    Schemas marked as draft placeholders where runtime bodies are not fully confirmed.
servers:
  - url: /api
    description: Legacy unversioned backend prefix (current)
security:
  - bearerAuth: []
...
paths:
...
  /did/system/init:
    post:
      tags: [DID]
      summary: Initialize system root entity
      operationId: didSystemInit
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /did/entity:
    post:
      tags: [DID]
      summary: Create sub-entity and issue VC
      operationId: didCreateEntity
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DidEntityRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /did/vp/generate:
    post:
      tags: [DID]
      summary: Generate verifiable presentation
      operationId: didVpGenerate
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Draft placeholder — holder_did, vcs, challenge
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /did/vc/reissue:
    post:
      tags: [DID]
      summary: Revoke and reissue VC
      operationId: didVcReissue
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DidCredentialRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /did/vc/revoke:
    post:
      tags: [DID]
      summary: Revoke VC
      description: |
        Revoke an issued verifiable credential. JWT required (SECURITY-A10, commit 898f99c).
        Request body is a draft placeholder — do not embed real VC, DID, or private key material in docs.
      operationId: didVcRevoke
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Draft placeholder — vc_id and revoked_by_did required at runtime
              properties:
                vc_id:
                  type: string
                  example: VC_ID_PLACEHOLDER
                revoked_by_did:
                  type: string
                  example: did:imbs:entity:placeholder
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/login:
    post:
      tags: [DID]
      summary: DID login
      description: Public — returns JWT. Supports VP or challenge/signature modes at runtime.
      operationId: didLogin
      security: []
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/me:
    get:
      tags: [DID]
      summary: Current authenticated DID entity
      operationId: didMe
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/challenge:
    get:
      tags: [DID]
      summary: Generate login challenge
      description: Public at runtime — no JWT.
      operationId: didChallenge
      security: []
      parameters:
        - name: length
          in: query
          required: false
          schema:
            type: integer
            default: 32
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/entity/{did}:
    get:
      tags: [DID]
      summary: Get entity by DID
      description: Public read at runtime — production auth hardening pending.
      operationId: didEntityGet
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/vp/verify:
    post:
      tags: [DID]
      summary: Verify verifiable presentation
      description: Public at runtime — no JWT decorator on Flask route.
      operationId: didVpVerify
      security: []
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/vc/status:
    post:
      tags: [DID]
      summary: Check verifiable credential status
      description: Public at runtime — no JWT decorator on Flask route.
      operationId: didVcStatus
      security: []
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/subordinates/direct:
    get:
      tags: [DID]
      summary: List direct subordinate entities
      description: Public read at runtime — requires parent_did query param.
      operationId: didSubordinatesDirect
      security: []
      parameters:
        - name: parent_did
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/subordinates/tree:
    get:
      tags: [DID]
      summary: Subordinate entity tree
      description: Public read at runtime — requires parent_did query param.
      operationId: didSubordinatesTree
      security: []
      parameters:
        - name: parent_did
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/subordinates/flat:
    get:
      tags: [DID]
      summary: Flat list of descendant entities
      description: Public read at runtime — requires parent_did query param.
      operationId: didSubordinatesFlat
      security: []
      parameters:
        - name: parent_did
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/chain/entity-hash/{did}:
    get:
      tags: [DID]
      summary: On-chain entity metadata hash
      description: Public read at runtime — blockchain contract must be configured.
      operationId: didChainEntityHash
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '503':
          description: Blockchain contract not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/chain/vc-hash/{vc_id}:
    get:
      tags: [DID]
      summary: On-chain VC hash
      description: Public read at runtime — blockchain contract must be configured.
      operationId: didChainVcHash
      security: []
      parameters:
        - name: vc_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '503':
          description: Blockchain contract not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/chain/events:
    get:
      tags: [DID]
      summary: Query blockchain contract events
      description: Public read at runtime — filter params vary; draft placeholder.
      operationId: didChainEvents
      security: []
      parameters:
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'
  /did/verify/entity/{did}:
    get:
      tags: [DID]
      summary: Verify entity on-chain
      description: Public read at runtime — verifies entity anchor status.
      operationId: didVerifyEntity
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XTraceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '500':
          $ref: '#/components/responses/ServerError'

  # --- System permissions (JWT) ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /api/did/login — use placeholder token in docs only
  parameters:
    XTraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Optional correlation id for audit (SECURITY-A9)
...
  schemas:
    ErrorResponse:
      type: object
      description: Existing Result.error style — draft
      properties:
        code:
          type: integer
        message:
          type: string
      additionalProperties: true
    SuccessResponse:
      type: object
      description: Existing Result.success style — draft
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      additionalProperties: true
...
    DidEntityRequest:
      type: object
      description: Draft placeholder — parent_did, entity_type_code, name, permissions required at runtime
      additionalProperties: true
      example:
        parent_did: did:imbs:system:root:placeholder
        entity_type_code: property
        name: Example Entity
        permissions:
          - device:read
    DidCredentialRequest:
      type: object
      description: Draft placeholder — issuer_did, subject_did required at runtime
      additionalProperties: true
  responses:
    Success:
      description: Success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    UnauthorizedError:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Invalid token
    ForbiddenError:
      description: |
        Forbidden — valid JWT but missing required permission (A6B/A7B/A10B),
        or environment guard denied (e.g. SSE test in production, A8B).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 403
            message: 'Permission denied: requires [modeling:train]'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'